Rework the arduino sketch
Go back to the interrupt based implementation, because the sampling implementation had too much error in the speed difference measurement.

Only major difference to original simpledyno sketch is now that the data is passed atomically to the main loop, i.e. timestamp and duration are now always from same sample.